% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare brms models}
\usage{
compare_models(x, criterion = "loo", ...)
}
\arguments{
\item{x}{A list of \code{brmsfit} objects that will be compared}

\item{criterion}{A character string indicating the criterion to use to compare models. For now, it can only take "loo" as value, which invokes Leave-One-Out Cross-Validation (LOO).}
}
\description{
Compare brms models
}
