% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_models.R
\name{fit_model}
\alias{fit_model}
\title{Specify and sample a brms model}
\usage{
fit_model(name, ...)
}
\arguments{
\item{name}{A character string indicating the name to be assigned to the model}

\item{...}{Arguments to be passed to \code{brms::brm()}}
}
\description{
Specify and sample a brms model
}
