% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preprocessing.R
\name{get_data}
\alias{get_data}
\title{Prepare data for analyses}
\usage{
get_data(items, responses, participants)
}
\arguments{
\item{items}{A data frame resulting from calling \code{get_items}}

\item{responses}{A data frame resulting from calling \code{get_responses}}

\item{participants}{A data frame resulting from calling \code{get_participants}}
}
\description{
Prepare data for analyses
}
