% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preprocessing.R
\name{get_items}
\alias{get_items}
\title{Get item data}
\usage{
get_items(multilex_data, class = c("Adjective", "Noun", "Verb"))
}
\arguments{
\item{multilex_data}{A named list resulting from calling \code{get_multilex}}

\item{class}{A character vector indicating the word classes to be included in the resulting dataset. Takes "Adjective", "Noun" and/or "Verb" as values.}
}
\description{
Get item data
}
