% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_importing.R
\name{get_multilex}
\alias{get_multilex}
\title{Get multilex data}
\usage{
get_multilex(update = FALSE, type = c("understands", "produces"))
}
\arguments{
\item{update}{A logical value. Should Multilex data be updated and new questionnaire responses be fetched?}

\item{type}{A character vector indicating for what type of measure should vocabulary and prevalence data be computed? It takes "understands" and/or "produces"}
}
\value{
A named list of data frames containing questionnaire responses, participant data, item data, prevalence data and vocabualary data
}
\description{
Get multilex data
}
