% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preprocessing.R
\name{get_participants}
\alias{get_participants}
\title{Get participant-level data}
\usage{
get_participants(
  multilex_data,
  longitudinal = "all",
  age = c(12, 32),
  lp = c("Monolingual", "Bilingual"),
  other_threshold = 0.1
)
}
\arguments{
\item{multilex_data}{A named list resulting from calling \code{get_multilex}}

\item{longitudinal}{Should longitudinal data be included? If "all" (default), all responses (including repeated measures) are included. If "no", participants with more than one responses to the questionnaire (regardless of the version) are excluded. If "first", only the first response of each participant is included. If "last", only the last response of each participant is included. If "only", only responses with repeated measures are included.}

\item{age}{Numeric vector of length two indicating the minimum and maximum age of participants that will be included in the resulting dataset.}

\item{lp}{Character vector indicating the language profile (LP) of the participants that will be included in the resulting dataset. In takes "Monolingual", "Bilingual", and/or "Other" as values.}

\item{other_threshold}{Numeric value between 0 and 1 indicating the minimum exposure to a language other than Catalan or Spanish that a participant need to be exposed to to be excluded.}
}
\description{
Get participant-level data
}
