% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preprocessing.R
\name{get_responses}
\alias{get_responses}
\title{Get item responses}
\usage{
get_responses(multilex_data, update = FALSE, longitudinal = "all")
}
\arguments{
\item{multilex_data}{A named list resulting from calling \code{get_multilex}}

\item{update}{A logical value. Should Multilex data be updated and new questionnaire responses be fetched?}

\item{longitudinal}{Should longitudinal data be included? If "all" (default), all responses (including repeated measures) are included. If "no", participants with more than one responses to the questionnaire (regardless of the version) are excluded. If "first", only the first response of each participant is included. If "last", only the last response of each participant is included. If "only", only responses with repeated measures are included.}
}
\description{
Get item responses
}
