% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_preprocessing.R
\name{get_vocabulary}
\alias{get_vocabulary}
\title{Get participant vocabulary data}
\usage{
get_vocabulary(
  multilex_data,
  age = c(12, 32),
  type = c("understands", "produces")
)
}
\arguments{
\item{multilex_data}{A named list resulting from calling \code{get_multilex}}

\item{age}{Numeric vector of length two indicating the minimum and maximum age of participants that will be included in the resulting dataset.}

\item{type}{A character vector indicating for what type of measure should vocabulary and prevalence data be computed? It takes "understands" and/or "produces".}
}
\description{
Get participant vocabulary data
}
